Implement weaving of the `Protected` attribute
In CIL aseemblies, protected members are marked with the `Family`
flag. This Fody weaver removes the `ProtectionAttributes.Protected`
custom attribute from all of the elements marked with it and sets
their `IsFamily` flags to true.

WARNING: this module DOES NOT check if the resulting file is
consistent. Please check it with `peverify`.